Update labels for best practices
If you nest input tags inside of the label tags you are linking them together. This allows users to click on the label (not only the box) and automatically focus onto the input box. Also, this format makes it easier for Screen Readers to identify forms to fill out on your form.

One other good alternative is to use is to use <label for=""> attribute, e.g.: 
<label for="firstname"></label>
<input id="firstname" name="firstname" type="text" value="<%= user ? user.get('firstname') : '' %>">